feat(web): unify collections and want-lists into a Binders library tab
"Collection" was carrying two meanings at once, so the Library kept collections (owned) and want-lists (chasing) in two disconnected tabs. This merges them under a single "Binders" tab — the shared physical metaphor a collector already uses — where each binder reads as one kind of thing: a palm "Owned" badge or a sun "Chasing" badge, with an All / Owned / Chasing filter to scope the list. Both kinds interleave newest-first.

Collections and want-lists stay separate API resources underneath; this is a presentation-layer bridge that composes useCollections + useWishlists. The owned-only smart-collection builder and both detail modals (SmartCollectionTarget for catalog targets, WishlistDetail for want-lists) move onto the new surface unchanged. Badge tones follow the OwnershipBadge convention (#576). Help copy and the user-scoped tab gate fold the two old tabs into one.

Closes #503
