+++ web/src/components/LibraryPanel.test.tsx
@@ -41,17 +41,16 @@ describe('LibraryPanel', () => {
     vi.restoreAllMocks()
   })
 
-  it('sidebar variant renders all four tabs with Searches active', async () => {
+  it('sidebar variant renders all three tabs with Searches active', async () => {
     render(<LibraryPanel variant="sidebar" onRun={vi.fn()} />)
-    // Collections / wishlists are gated on a resolved signed-in user,
-    // so wait for the auth load before asserting the full tab set.
-    expect(await screen.findByRole('tab', { name: /Collections/i })).toBeInTheDocument()
+    // Binders is gated on a resolved signed-in user, so wait for the auth
+    // load before asserting the full tab set.
+    expect(await screen.findByRole('tab', { name: /Binders/i })).toBeInTheDocument()
     expect(screen.getByRole('tab', { name: /Searches/i })).toHaveAttribute(
       'aria-selected',
       'true',
     )
     expect(screen.getByRole('tab', { name: /Recent/i })).toBeInTheDocument()
-    expect(screen.getByRole('tab', { name: /Wishlists/i })).toBeInTheDocument()
     expect(await screen.findByText(/No saved searches yet/i)).toBeInTheDocument()
   })
 
@@ -70,19 +69,11 @@ describe('LibraryPanel', () => {
     expect(screen.getByText('Pikachu')).toBeInTheDocument()
   })
 
-  it('clicking the Collections tab fetches and shows the empty state', async () => {
+  it('clicking the Binders tab fetches and shows the empty state', async () => {
     render(<LibraryPanel variant="sidebar" onRun={vi.fn()} />)
-    fireEvent.click(await screen.findByRole('tab', { name: /Collections/i }))
+    fireEvent.click(await screen.findByRole('tab', { name: /Binders/i }))
     expect(
-      await screen.findByText(/You don't have any collections yet/i),
-    ).toBeInTheDocument()
-  })
-
-  it('clicking the Wishlists tab fetches and shows the empty state', async () => {
-    render(<LibraryPanel variant="sidebar" onRun={vi.fn()} />)
-    fireEvent.click(await screen.findByRole('tab', { name: /Wishlists/i }))
-    expect(
-      await screen.findByText(/You don't have any wishlists yet/i),
+      await screen.findByText(/You don't have any binders yet/i),
     ).toBeInTheDocument()
   })
 
@@ -112,7 +103,7 @@ describe('LibraryPanel', () => {
     expect(screen.getByRole('tab', { name: /Searches/i })).toBeInTheDocument()
   })
 
-  it('hides the Collections + Wishlists tabs when no user is identified', async () => {
+  it('hides the Binders tab when no user is identified', async () => {
     vi.mocked(client.fetchMe).mockResolvedValue({ user: null, authEnabled: true })
     render(<LibraryPanel variant="sidebar" onRun={vi.fn()} />)
 
@@ -121,19 +112,17 @@ describe('LibraryPanel', () => {
     ).toBeInTheDocument()
     expect(screen.getByRole('tab', { name: /Searches/i })).toBeInTheDocument()
     expect(screen.getByRole('tab', { name: /Recent/i })).toBeInTheDocument()
-    expect(screen.queryByRole('tab', { name: /Collections/i })).not.toBeInTheDocument()
-    expect(screen.queryByRole('tab', { name: /Wishlists/i })).not.toBeInTheDocument()
+    expect(screen.queryByRole('tab', { name: /Binders/i })).not.toBeInTheDocument()
   })
 
-  it('shows the Collections + Wishlists tabs on self-host (authEnabled false)', async () => {
+  it('shows the Binders tab on self-host (authEnabled false)', async () => {
     vi.mocked(client.fetchMe).mockResolvedValue({
       user: { id: 1, email: null, display_name: 'default' },
       authEnabled: false,
     })
     render(<LibraryPanel variant="sidebar" onRun={vi.fn()} />)
 
-    expect(await screen.findByRole('tab', { name: /Collections/i })).toBeInTheDocument()
-    expect(screen.getByRole('tab', { name: /Wishlists/i })).toBeInTheDocument()
+    expect(await screen.findByRole('tab', { name: /Binders/i })).toBeInTheDocument()
   })
 
   it('does not leak a stale saved-search count when no user is identified', async () => {
